% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_data.R
\name{desc_data}
\alias{desc_data}
\title{Describe rectangular data}
\usage{
desc_data(data, metadata, bullet = "-", dbl_space = FALSE)
}
\arguments{
\item{data}{Dataframe}

\item{metadata}{Dataframe with two columns:
\itemize{
\item "col" (column names in \code{data})
\item "desc" (description of each column).
}

"col" must match column names in \code{data} exactly (including order)}

\item{bullet}{Character vector of length 1; type of bullet to print before
each line of output. May choose from '-', '*', or 'none'.}

\item{dbl_space}{Logical vector of length 1; should the output have double
spacing? (an extra line break between each line of text)}
}
\value{
Character string formatted for including in data README
}
\description{
Intended use is for creating a description of rectangular data in a data
README file.
}
\examples{
cars_small <- mtcars[1:10, 1:3]
cars_metadata <- data.frame(
  col = c("mpg", "cyl", "disp"),
  desc = c("Miles per gallon", "Number of cylinders", "Displacement")
)
desc_data(cars_small, cars_metadata)
}
