% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain_document.R
\name{plain_document}
\alias{plain_document}
\title{Convert to a plain text document}
\usage{
plain_document(wrap = "auto", columns = 72)
}
\arguments{
\item{wrap}{Character vector of length 1; type of text wrapping
to use. Must choose one of "auto", "none", or "preserve".}

\item{columns}{Numeric vector of length 1; number of characters
to write per line.}
}
\value{
R Markdown output format to pass to rmarkdown::render()
}
\description{
Format for converting from R Markdown to a plain text document.
}
\details{
As per pandoc documentation for
\href{https://pandoc.org/MANUAL.html#option--wrap}{\code{wrap}}, "With 'auto' (the
default), pandoc will attempt to wrap lines to the column width specified by
\code{columns} (default 72). With 'none', pandoc will not wrap lines at all. With
'preserve', pandoc will attempt to preserve the wrapping from the source
document (that is, where there are nonsemantic newlines in the source, there
will be nonsemantic newlines in the output as well)."
}
\examples{
# Write out an example Rmd file
rmd_tempfile <- tempfile(fileext = ".Rmd")
lines <- c(
  "---",
  "title: 'Report'",
  "output_format: plain_document",
  "---",
  "This is a **markdown document**. But the formatting is *gone*",
  "when rendered to _plain text_."
)
write(lines, rmd_tempfile)

# Render to temporary text file
txt_tempfile <- tempfile(fileext = ".txt")
rmarkdown::render(
  rmd_tempfile, output_file = txt_tempfile,
  output_format = plain_document(columns = 40))

# Check output
readLines(txt_tempfile)

# Delete temporary files
unlink(rmd_tempfile)
unlink(txt_tempfile)

}
